:hammer: fix: state 초기값 krDate로 설정하고 loading 삭제
